Moved bitbucket webhook into its own file bitbucket.py
