+++ zproject/urls.py
@@ -159,7 +159,7 @@ urlpatterns += patterns('zerver.views',
     url(r'^api/v1/external/jira$',          'webhooks.jira.api_jira_webhook'),
     url(r'^api/v1/external/pivotal$',       'webhooks.pivotal.api_pivotal_webhook'),
     url(r'^api/v1/external/newrelic$',      'webhooks.newrelic.api_newrelic_webhook'),
-    url(r'^api/v1/external/bitbucket$',     'webhooks.api_bitbucket_webhook'),
+    url(r'^api/v1/external/bitbucket$',     'webhooks.bitbucket.api_bitbucket_webhook'),
     url(r'^api/v1/external/desk$',          'webhooks.api_deskdotcom_webhook'),
     url(r'^api/v1/external/stash$',         'webhooks.api_stash_webhook'),
     url(r'^api/v1/external/freshdesk$',     'webhooks.api_freshdesk_webhook'),
